+++ prestaspot/blocks/product-list/index.asset.php
@@ -9,5 +9,5 @@ return array(
         'wp-i18n',
         'wp-api-fetch',
     ),
-    'version' => '0.13.0',
+    'version' => '0.14.0',
 );