+++ prestaspot/includes/class-presta-spot-admin.php
@@ -123,6 +123,12 @@ class Presta_Spot_Admin
             'default' => Presta_Spot_Settings::PRESTASPOT_SETTINGS_DEFAULTS[Presta_Spot_Settings::SHOW_PRICE],
         ));
 
+        register_setting('prestaspot_settings_group', 'prestaspot_show_stock_status', array(
+            'type' => 'boolean',
+            'sanitize_callback' => array($this, 'sanitize_boolean'),
+            'default' => Presta_Spot_Settings::PRESTASPOT_SETTINGS_DEFAULTS[Presta_Spot_Settings::SHOW_STOCK_STATUS],
+        ));
+
         register_setting('prestaspot_settings_group', 'prestaspot_price_position', array(
             'type' => 'string',
             'sanitize_callback' => array($this, 'sanitize_price_position'),